Rudimentary check of extended character classes.
If we find one of the potentially-offending classes in an extended
bracketed character class we check to see if it is intersected with
something that forces ASCII or some relevant subset thereof.

The check is pretty simplistic: just see if there is an adjacent '&'
with one of the sanitizers on the other side of it.