use() in exampleas/numeric-character-classes
The cryptic title is my attempt to shoehorn into Git's 50-character
limit the fact that examples/numeric-character-classes now explicitly
loads ::RegularExpressions::ProhibitNumericCharacterClasses.

This is superfluous as far as Perl::Critic is concerned, but is useful
if you want to run a debugger without either explicitly loading the
class in the command line or trying to figure out where to set a
breakpoint after the policy is loaded so you can then set a breakpoint
in the policy.