Config items to allow single-script classes
If the allow_if_single_script Boolean item is true, \d and [:digit:] are
allowed within the scope of (*script_run: ... ) or equivalent.

If the allow_if_singleton Boolean item is true, \d and [:digit:] are
allowed if they are unquantified or quantified to 1 or less, AND if they
are not followed by another \d or [:digit:]

Collateral with this, I decided not to simply ban \d and [:digit:] in
old-style bracketed character classes, but accept them if the character
class also contains non-digit stuff.