examples/numeric-character-classes verbatim code
The above is my attempt to cram into Git's 50-character limit the fact
that examples/numeric-character-classes now takes verbatim code on the
command line as well as file names. If something that does not look like
an option also does not represent an extant file, it is written to a
temporary file which is passed to Perl::Critic.